+++ Source/Scene/ModelMesh.js
@@ -1,5 +1,8 @@
 /*global define*/
-define(function() {
+define([
+        '../Core/defineProperties'
+    ], function(
+        defineProperties) {
     "use strict";
 
     /**
@@ -13,18 +16,46 @@ define(function() {
      *
      * @see Model#getMesh
      */
-    var ModelMesh = function(name) {
+    var ModelMesh = function(mesh, id) {
+        this._name = mesh.name;
+        this._id = id;
+    };
+
+    defineProperties(ModelMesh.prototype, {
         /**
          * The value of the <code>name</code> property of this mesh.  This is the
          * name assigned by the artist when the asset is created.  This can be
-         * different than the name of the mesh property, which is internal to glTF.
+         * different than the name of the mesh property ({@link ModelMesh#id}),
+         * which is internal to glTF.
+         *
+         * @memberof ModelMesh.prototype
          *
          * @type {String}
-         *
          * @readonly
          */
-        this.name = name;
-    };
+        name : {
+            get : function() {
+                return this._name;
+            }
+        },
+
+        /**
+         * The name of the glTF JSON property for this mesh.  This is guaranteed
+         * to be unique among all meshes.  It may not match the mesh's <code>
+         * name</code> property (@link ModelMesh#name), which is assigned by
+         * the artist when the asset is created.
+         *
+         * @memberof ModelMesh.prototype
+         *
+         * @type {String}
+         * @readonly
+         */
+        id : {
+            get : function() {
+                return this._id;
+            }
+        }
+    });
 
     return ModelMesh;
 });